+++ typeutils/datatype.go
@@ -1,9 +1,11 @@
 package typeutils
 
 import (
+	"fmt"
 	"reflect"
 
 	"github.com/piyushsingariya/shift/types"
+	"github.com/piyushsingariya/shift/utils"
 )
 
 // TypeFromValue return DataType from v type
@@ -34,5 +36,43 @@ func TypeFromValue(v interface{}) types.DataType {
 		return types.OBJECT
 	default:
 		return types.UNKNOWN
+	}
+}
+
+func MaximumOnDataType[T any](typ types.DataType, a, b T) (T, error) {
+	switch typ {
+	case types.TIMESTAMP:
+		adate, err := ReformatDate(a)
+		if err != nil {
+			return a, fmt.Errorf("failed to reformat[%v] while comparing: %s", a, err)
+		}
+		bdate, err := ReformatDate(b)
+		if err != nil {
+			return a, fmt.Errorf("failed to reformat[%v] while comparing: %s", b, err)
+		}
+
+		if utils.MaxDate(adate, bdate) == adate {
+			return a, nil
+		}
+
+		return b, nil
+	case types.INT64:
+		aint, err := ReformatInt64(a)
+		if err != nil {
+			return a, fmt.Errorf("failed to reformat[%v] while comparing: %s", a, err)
+		}
+
+		bint, err := ReformatInt64(b)
+		if err != nil {
+			return a, fmt.Errorf("failed to reformat[%v] while comparing: %s", b, err)
+		}
+
+		if aint > bint {
+			return a, nil
+		}
+
+		return b, nil
+	default:
+		return a, fmt.Errorf("comparison not available for data types %v now", typ)
 	}
 }
